refactor: clean up mysql restore package
